Added support for disconnection
1. When connection is impossible, paramiko raises a subclass of
socket.error; this is caught in the queue_managing py-script - which
sets all cluster stats to 0, and percentage_active to None.

2. When percentage_active of a host is None, a spanning entry is added
to the top of the job list, stating "Couldn't connect to server!"

3. base.html was occasionally unable to render an accordion collapse,
as jquery.js couldn't be accessed. This left a frozen page. Made
jquery.min.js a static object to fix this.